+++ eslint.config.ts
@@ -2,20 +2,22 @@ import reactHooksPlugin from "eslint-plugin-react-hooks";
 import reactRefreshPlugin from "eslint-plugin-react-refresh";
 import { globalIgnores } from "eslint/config";
 import globals from "globals";
-import tseslint from "typescript-eslint";
+import tseslint, { parser } from "typescript-eslint";
 
-export default tseslint.config([
-  globalIgnores(["coverage", "dist", "src/**/*.d.ts"]),
+export default [
+  globalIgnores(["build", "coverage", "dist", "src/**/*.d.ts", "*.config.ts"]),
+  ...tseslint.configs.recommended,
+  reactHooksPlugin.configs["recommended-latest"],
+  reactRefreshPlugin.configs.vite,
   {
     files: ["**/*.ts", "**/*.tsx"],
-    extends: [
-      tseslint.configs.recommended,
-      reactHooksPlugin.configs["recommended-latest"],
-      reactRefreshPlugin.configs.vite,
-    ],
     languageOptions: {
       ecmaVersion: 2020,
       globals: globals.browser,
+      parserOptions: {
+        project: ["tsconfig.json"],
+        tsconfigRootDir: __dirname,
+      },
     },
   },
-]);
+];